Initial attempt at migrating from bootstrap 4 to 5
This work will need the 629-upgrade-bootstrap-5 MATScommon branch checked out as
well. In MATS - I've upgraded bootstrap to v5 everywhere I could find it and
removed the duplicate bootstrap 3 dependency from the atmosphere
package list. I'm not sure if this was wise. Either way, it resulted in a very broken
MATS UI.