Let the meteor bundle be owned by root
This is a best practice so the app can't rewrite its own binaries. We do
need to be able to write to the settings file and the node-file-cache
location so the Dockerfile updates the permissions for those locations.

Since these locations will be owned by root - create them in the
Dockerfile and not in the app launcher script.