Force a rebuild of the mats-common dependencies
The mats-common Meteor Atmosphere package installs some native
dependencies in the first stage of this Dockerfile. That stage uses a
Ubuntu base image so causes problems when we switch to the
musl-based Alpine production image. We need to 'npm rebuild' in the
meteor package's node_modules as well as the places the
'build-meteor-npm-dependencies.sh' script does it.