Split the Environment config into two options
The "Environment" config option had two effects - it set the subtitle on
the page if it was anything other than "Production" and also would
show/hide the toggle for experimental GSL apps if it was set to
"Production". This overloading was undesired and so I've created a new
"EnvironmentLabel" option to control the subtitle in the UI and a
"IsProduction" option to toggle show/hide behavior.